\name{lmfor-package}
\alias{lmfor-package}
\alias{lmfor}
\docType{package}
\title{
Functions of Lauri Mehtatalo
}
\description{
Functions for different purposes related to Forest biometrics, 
including illustrative graphics, numerical computation, modeling height-diameter relationships, 
prediction of tree volumes, modeling of diameter distributions and estimation off stand density using ITD. S
everal empirical datasets are also included.
}
\details{
\tabular{ll}{
Package: \tab lmfor\cr
Type: \tab Package\cr
Version: \tab 1.3\cr
Date: \tab 2019-12-03\cr
License: \tab GPL -2 \cr
LazyLoad: \tab yes\cr
}

}
\author{
Lauri Mehtatalo <lauri.mehtatalo@uef.fi>
}
\references{
Mehtatalo, L. and Lappi, J. forthcoming. Forest Biometrics with examples in R. 
Textbook in preparation.
Available at \url{http://cs.uef.fi/~lamehtat/}.
}

\keyword{ forest biometrics }
