\name{lmfor-package}
\alias{lmfor-package}
\alias{lmfor}
\docType{package}
\title{
Functions of Lauri Mehtatalo
}
\description{
Functions for different purposes related to Forest biometrics, including illustrative graphics, 
numerical computation, modeling height-diameter relationships, prediction of tree volumes, 
and modelling of diameter distributions. Datasets on tree height-diameter relationship,  
light response of moss photosynthesis, and productivity of stump-lifting machines are included.
}
\details{
\tabular{ll}{
Package: \tab lmfor\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2012-25-05\cr
License: \tab GPL -2 \cr
LazyLoad: \tab yes\cr
}

}
\author{
Lauri Mehtatalo <lauri.mehtatalo@uef.fi>
}
\references{
Mehtatalo, L. 2008. Forest Biometrics with examples in R. 
Lecture notes for the forest biometrics course. 186 pages. 
Available at \url{http://cs.uef.fi/~lamehtat/}.
}

\keyword{ forest biometrics }
