\name{lmfor-package}
\alias{lmfor-package}
\alias{lmfor}
\docType{package}
\title{
Functions of Lauri Mehtatalo
}
\description{
Functions for different purposes related to Forest biometrics, 
including illustrative graphics, numerical computation, modeling height-diameter relationships, 
prediction of tree volumes, modeling of diameter distributions and estimation off stand density using ITD. 
Several empirical datasets are also included; those data sets are used in the examples of Mehtatalo and Lappi (2020a, 2020b).
}
\details{
\tabular{ll}{
Package: \tab lmfor\cr
Type: \tab Package\cr
Version: \tab 1.5\cr
Date: \tab 2020-06-16\cr
License: \tab GPL -2 \cr
LazyLoad: \tab yes\cr
}

}
\author{
Lauri Mehtatalo <\email{lauri.mehtatalo@uef.fi}> and Kasper Kansanen <\email{kasperkansanen@gmail.com}>
}
\references{
Mehtatalo, Lauri and Lappi, Juha 2020a. Biometry for Forestry and Environmental Data: with examples in R. New York: Chapman and Hall/CRC. 426 p. \doi{10.1201/9780429173462}.

Mehtatalo, Lauri and Lappi, Juha 2020b. Biometry for Forestry and Environmental Data: with examples in R. Full Versions of The Web Examples. Available at \url{http://www.biombook.org}. 
}

\keyword{ forest biometrics }
